Moved GameBoard class into its own file.
